requirements.txt
numpy==1.26.4
fastapi==0.115.0
uvicorn[standard]==0.30.1
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
qdrant-client==1.8.2
python-multipart==0.0.9
requests==2.31.0
